requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
langchain-community
langchain-text-splitters
chromadb
pymupdf
python-multipart
pydantic-settings
python-dotenv
python-jose[cryptography]

# Frontend
streamlit

# Testing
pytest
httpx
